Fix types

+++ backend/dotnet/VY.Hackathon.TeamOne/VY.Hackaton.TeamOne.ProviderData.Infrastructure/Contracts/Models/ProviderDataRequest.cs
@@ -6,9 +6,9 @@
     );
 
     public record TimeData(
-        int Jardineria,
+        double Jardineria,
         double Equipaje,
-        int Coordination
+        double Coordination
     );
 
     public record HandlingFunctionHourPrice(
